Coding/GenAI/PoC/1 - DOCUMENT COMPARISON/49-RegulatoryDocsCompare/requirements.txt
streamlit
python-docx
diff-match-patch
pandas 
#difflib
ollama